Deler opp SearchConcept for mer oversiktlig kode
