+++ src/containers/SearchPage/components/results/SearchConcept/FormView.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const FormView = () => {
+  return <div>form</div>;
+};
+
+export default FormView;
