+++ src/containers/SearchPage/components/results/SearchConcept/index.js
@@ -1,0 +1,3 @@
+import SearchConcept from './SearchConcept';
+
+export default SearchConcept;
